refactor(cli): fetch_pr_*_via_gh + edit_or_create_repo_guard を cli/pr.rs へ移行 (SPEC-1942)
SPEC-1942 SC-025 進捗の継続。 cli.rs から PR family 関連の gh 呼び出し
helpers を全て cli/pr.rs へ移動 (pub(super))。

移動対象 (18 helpers):
- fetch_current_pr_via_gh / create_pr_via_gh / edit_pr_via_gh
- comment_on_pr_via_gh / fetch_pr_reviews_via_gh
- fetch_pr_review_threads_via_gh / reply_and_resolve_pr_review_threads_via_gh
- fetch_pr_checks_via_gh / fetch_pr_review_thread_state_via_gh
- review_thread_has_comment_body / should_reply_to_review_thread
- should_resolve_review_thread / parse_pr_checks_items_json
- parse_pr_checks_items_response / parse_available_fields
- extract_pr_url / parse_pr_number_from_url / edit_or_create_repo_guard

cli.rs 2572 → 1992 行 (-580)、 pr.rs 397 → 994 行 (+597)。 env.rs 内の
super::fn_name 呼び出しは super::pr::fn_name に更新。 cli.rs 内の test
は qualified path (crate::cli::pr::*) で参照を維持。 機能・契約・test
結果は不変 (lib 345 / cli_test 68 全 pass、 clippy 0 / fmt clean)。
